Refactor functions into own respective files
