+++ src/_tests_/first.test.js
@@ -1,0 +1,7 @@
+import { first } from 'first'
+
+describe('first()', () => {
+  it('should return first element of an array', () => {
+    expect(first([ 'Heinrich Kerze', 'Jonny Ranch' ])).toBe('Heinrich Kerze')
+  })
+}) // end first()
