+++ src/_tests_/generateMarkup.test.js
@@ -1,0 +1,9 @@
+import { generateMarkup } from '../generateMarkup'
+
+describe('generateMarkup()', () => {
+  it('should return wrapped object', () => {
+    expect(generateMarkup('myHtml')).toEqual({
+      __html: 'myHtml'
+    })
+  })
+}) // end generateMarkup()
