+++ src/_tests_/last.test.js
@@ -1,0 +1,8 @@
+import { last } from 'last'
+
+describe('last()', () => {
+  it('should return last element of an array', () => {
+    expect(last([ 'Heinrich Kerze', 'Zylwira Zergler', 'Jonny Ranch' ]))
+      .toBe('Jonny Ranch')
+  })
+}) // end last()
