+++ src/first.js
@@ -1,0 +1,9 @@
+// @flow
+
+/**
+  * Returns the first element of an array
+  * @function first
+  * @param {array} arr - Array
+  * @return {element} First element of the array
+  */
+export const first = <T>(arr: Array<T>): T => arr[0]
